# internal/user/entity/user.go
// Package entity defines domain entities for the user module.
package entity

import "time"

// User represents a user in the system with profile information.
type User struct {
	ID        int64     `json:"id"`
	Username  string    `json:"username"`
	Email     string    `json:"email"`
	Bio       string    `json:"bio"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}